#5 Implement event consumer

+++ domain/src/main/java/com/antharos/analytics/domain/DepartmentSalary.java
@@ -1,5 +1,6 @@
 package com.antharos.analytics.domain;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record DepartmentSalary(UUID departmentId, long totalSalary) {}
+public record DepartmentSalary(UUID departmentId, BigDecimal totalSalary) {}
