+++ domain/src/main/java/com/antharos/analytics/domain/MonthlySalaryCost.java
@@ -1,5 +1,6 @@
 package com.antharos.analytics.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record MonthlySalaryCost(LocalDate month, long totalSalary) {}
+public record MonthlySalaryCost(LocalDate month, BigDecimal totalSalary) {}
